requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.12
chromadb==0.4.22
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
tiktoken==0.5.2
streamlit==1.29.0